requirements.txt
# Core FastAPI dependencies
fastapi==0.115.13
uvicorn[standard]==0.32.1
websockets==12.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1

# Utility dependencies
pytz==2023.3
psutil==5.9.6
structlog==24.1.0
langdetect==1.0.9
redis==5.2.0
# pybreaker==1.3.0  # Removed due to Python 3.12 compatibility issues
dependency-injector>=4.41.0
slowapi==0.1.9
aiofiles==23.2.1

# Security dependencies
PyJWT==2.8.0
passlib==1.7.4
email-validator==2.1.0
bcrypt==4.3.0
cryptography>=42.0.0

# OCR and image processing
pytesseract==0.3.10
Pillow>=10.1.0
PyMuPDF==1.24.1
opencv-python==4.8.1.78
easyocr==1.7.2

# ML and AI dependencies - FIXED FOR PYTHON 3.12 COMPATIBILITY
langchain==0.0.350
langchain-community==0.0.10
numpy>=1.23.5,<2.0.0  # Compatible with langchain and Python 3.12
scipy==1.15.3
scikit-learn==1.7.0
sentence-transformers==2.2.2
faiss-cpu==1.11.0  # CPU version - use requirements-gpu.txt for GPU support
ollama==0.5.1

# OpenTelemetry packages for monitoring and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-jaeger==1.21.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Additional dependencies
celery==5.3.4
kombu==5.3.4
aiohttp>=3.8.0
memory_profiler>=0.61.0
psycopg2-binary>=2.9.0
boto3>=1.35.0
botocore>=1.35.0

# Data processing and normalization
pandas>=2.0.0
flashtext>=2.7
rapidfuzz>=3.0.0
jsonschema>=4.21.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
asgi_lifespan==2.1.0

# Added for unstructured data processing
unstructured
python-magic
lxml
beautifulsoup4
html5lib
markdown
tabulate
openpyxl
python-docx
EbookLib
msg-parser
onedrivesdk
pypandoc
python-pptx
python-dateutil

# Google API dependencies for Gmail Inbox Zero
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0